Runtime/r3/fileio.cpp: Make the parameters of RTFileSetForceFlags() 64-bit like all the other RTFile* APIs for the mode flags, bugref:3409
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@107803 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/iprt/file.h
@@ -297,7 +297,7 @@ typedef RTFILEACTION *PRTFILEACTION;
  * @param   fSet            Open flags to be forced set.
  * @param   fMask           Open flags to be masked out.
  */
-RTR3DECL(int)  RTFileSetForceFlags(unsigned fOpenForAccess, unsigned fSet, unsigned fMask);
+RTR3DECL(int)  RTFileSetForceFlags(uint64_t fOpenForAccess, uint64_t fSet, uint64_t fMask);
 #endif /* IN_RING3 */
 
 /**
