+++ src/VBox/Runtime/r3/fileio.cpp
@@ -53,22 +53,22 @@
 *   Global Variables                                                                                                             *
 *********************************************************************************************************************************/
 /** Set of forced set open flags for files opened read-only. */
-static unsigned g_fOpenReadSet = 0;
+static uint64_t g_fOpenReadSet = 0;
 
 /** Set of forced cleared open flags for files opened read-only. */
-static unsigned g_fOpenReadMask = 0;
+static uint64_t g_fOpenReadMask = 0;
 
 /** Set of forced set open flags for files opened write-only. */
-static unsigned g_fOpenWriteSet = 0;
+static uint64_t g_fOpenWriteSet = 0;
 
 /** Set of forced cleared open flags for files opened write-only. */
-static unsigned g_fOpenWriteMask = 0;
+static uint64_t g_fOpenWriteMask = 0;
 
 /** Set of forced set open flags for files opened read-write. */
-static unsigned g_fOpenReadWriteSet = 0;
+static uint64_t g_fOpenReadWriteSet = 0;
 
 /** Set of forced cleared open flags for files opened read-write. */
-static unsigned g_fOpenReadWriteMask = 0;
+static uint64_t g_fOpenReadWriteMask = 0;
 
 
 /**
@@ -80,7 +80,7 @@ static unsigned g_fOpenReadWriteMask = 0;
  * @param   fSet            Open flags to be forced set.
  * @param   fMask           Open flags to be masked out.
  */
-RTR3DECL(int)  RTFileSetForceFlags(unsigned fOpenForAccess, unsigned fSet, unsigned fMask)
+RTR3DECL(int)  RTFileSetForceFlags(uint64_t fOpenForAccess, uint64_t fSet, uint64_t fMask)
 {
     /*
      * For now allow only RTFILE_O_WRITE_THROUGH. The other flags either
